requirements.txt
Django==6.0.1
djangorestframework
django-cors-headers
python-dotenv
requests
fastapi
uvicorn
pydantic
numpy
